@@ -2,15 +2,15 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <RootNamespace>net.jommy.RuuviCore</RootNamespace>
     <AssemblyName>ruuvi-core</AssemblyName>
-    <Version>0.6.0</Version>
+    <Version>1.0.0</Version>
     <Authors>Jomi Laakkonen</Authors>
     <Company />
-    <PackageVersion>0.7.0</PackageVersion>
-    <AssemblyVersion>0.7.1</AssemblyVersion>
-    <FileVersion>0.7.1</FileVersion>
+    <PackageVersion>1.0.0</PackageVersion>
+    <AssemblyVersion>1.0.0</AssemblyVersion>
+    <FileVersion>1.0.0</FileVersion>
     <UserSecretsId>d40304c8-3507-41b4-883a-dca7cf14316c</UserSecretsId>
     <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
     <StartupObject>net.jommy.RuuviCore.Program</StartupObject>
@@ -22,15 +22,14 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Append.Blazor.Notifications" Version="1.1.0" />
-    <PackageReference Include="BlazorStrap" Version="5.0.106" />
-    <PackageReference Include="Microsoft.Bcl.AsyncInterfaces" Version="6.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="6.0.1" />
-    <PackageReference Include="Microsoft.Orleans.Server" Version="3.6.5" />
-    <PackageReference Include="Serilog" Version="2.12.0" />
-    <PackageReference Include="Serilog.Extensions.Hosting" Version="5.0.1" />
-    <PackageReference Include="Serilog.Extensions.Logging" Version="3.1.0" />
-    <PackageReference Include="Serilog.Sinks.Console" Version="4.1.0" />
+    <PackageReference Include="Microsoft.Bcl.AsyncInterfaces" Version="8.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Hosting" Version="8.0.0" />
+    <PackageReference Include="Microsoft.Orleans.Serialization.SystemTextJson" Version="8.0.0-rc2" />
+    <PackageReference Include="Microsoft.Orleans.Server" Version="8.0.0-rc2" />
+    <PackageReference Include="Serilog" Version="3.1.1" />
+    <PackageReference Include="Serilog.Extensions.Hosting" Version="8.0.0" />
+    <PackageReference Include="Serilog.Extensions.Logging" Version="8.0.0" />
+    <PackageReference Include="Serilog.Sinks.Console" Version="5.0.1" />
     <PackageReference Include="Serilog.Sinks.File" Version="5.0.0" />
   </ItemGroup>
 
@@ -43,12 +42,9 @@
   </ItemGroup>
   
   <ItemGroup>
-    <None Update="appsettings.core.json">
+    <None Update="appsettings.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
   </ItemGroup>
-  
-  <ItemGroup>
-    <Folder Include="Services" />
-  </ItemGroup>
+
 </Project>
