@@ -1,14 +1,15 @@
-namespace net.jommy.RuuviCore.Common
-{
-    public class DBusSettings
-    {
-        private const string DefaultAdapterName = "hci0";
-        
-        public string BluetoothAdapterName { get; set; }
+namespace net.jommy.RuuviCore.Common;
 
-        public DBusSettings()
-        {
-            BluetoothAdapterName = DefaultAdapterName;
-        }
+public class DBusSettings
+{
+    private const string DefaultAdapterName = "hci0";
+        
+    public string BluetoothAdapterName { get; set; }
+
+    public bool Simulate { get; set; }
+
+    public DBusSettings()
+    {
+        BluetoothAdapterName = DefaultAdapterName;
     }
 }
