@@ -1,7 +1,9 @@
-﻿namespace net.jommy.Orleans
+﻿namespace net.jommy.Orleans;
+
+public class FileStorageProviderOptions
 {
-    public class FileStorageProviderOptions
-    {
-        public string Directory { get; set; }
-    }
+    public string Directory { get; set; }
+    public string ServiceId { get; set; }
+
+    public string Name { get; set; }
 }
