@@ -1,28 +1,32 @@
 ﻿using System;
 
-namespace net.jommy.RuuviCore.Interfaces
+namespace net.jommy.RuuviCore.Interfaces;
+
+[Serializable]
+[Orleans.GenerateSerializer]
+[Orleans.Alias("Acceleration")]
+public class Acceleration
 {
-    [Serializable]
-    public class Acceleration
+    [Orleans.Id(0)]
+    public decimal XAxis { get; set; }
+    [Orleans.Id(1)]
+    public decimal YAxis { get; set; }
+    [Orleans.Id(2)]
+    public decimal ZAxis { get; set; }
+
+    public Acceleration()
     {
-        public decimal XAxis { get; set; }
-        public decimal YAxis { get; set; }
-        public decimal ZAxis { get; set; }
+    }
 
-        public Acceleration()
-        {
-        }
+    public Acceleration(decimal xAxis, decimal yAxis, decimal zAxis)
+    {
+        XAxis = xAxis;
+        YAxis = yAxis;
+        ZAxis = zAxis;
+    }
 
-        public Acceleration(decimal xAxis, decimal yAxis, decimal zAxis)
-        {
-            XAxis = xAxis;
-            YAxis = yAxis;
-            ZAxis = zAxis;
-        }
-
-        public override string ToString()
-        {
-            return $"X:{XAxis}, Y:{YAxis}, Z:{ZAxis}";
-        }
+    public override string ToString()
+    {
+        return $"X:{XAxis}, Y:{YAxis}, Z:{ZAxis}";
     }
 }
