@@ -3,14 +3,13 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Orleans;
 
-namespace net.jommy.RuuviCore.Interfaces
-{
-    public interface IGrainRegistry<T> : IGrainWithIntegerKey
-    {
-        public const int GrainRegistryId = 0;
-        
-        Task<List<T>> GetRegisteredGrains();
+namespace net.jommy.RuuviCore.Interfaces;
 
-        Task RegisterInfo(string id, string name, DateTime? lastSeen);
-    }
+public interface IGrainRegistry<T> : IGrainWithIntegerKey
+{
+    public const int GrainRegistryId = 0;
+        
+    Task<List<T>> GetRegisteredGrains();
+
+    Task RegisterInfo(string id, string name, DateTime? lastSeen);
 }
