@@ -2,19 +2,37 @@ using System;
 using System.Collections.Generic;
 using net.jommy.RuuviCore.Common;
 
-namespace net.jommy.RuuviCore.Interfaces
+namespace net.jommy.RuuviCore.Interfaces;
+
+[Serializable]
+[Orleans.GenerateSerializer]
+[Orleans.Alias("RuuviTag")]
+public class RuuviTag
 {
-    [Serializable]
-    public class RuuviTag
-    {
-        public string MacAddress { get; set; }
-        public string Name { get; set; }
-        public int DataSavingInterval { get; set; }
-        public bool StoreAcceleration { get; set; }
-        public bool StoreName { get; set; }
-        public bool DiscardMinMaxValues { get; set; }
-        public bool AllowMeasurementsThroughGateway { get; set; }
-        public bool IncludeInDashboard { get; set; }
-        public IDictionary<string, AlertThresholds> AlertRules { get; set; }
-    }
+    [Orleans.Id(0)]
+    public string MacAddress { get; set; }
+ 
+    [Orleans.Id(1)]
+    public string Name { get; set; }
+
+    [Orleans.Id(2)]
+    public int DataSavingInterval { get; set; }
+    
+    [Orleans.Id(3)]
+    public bool StoreAcceleration { get; set; }
+    
+    [Orleans.Id(4)]
+    public bool StoreName { get; set; }
+    
+    [Orleans.Id(5)]
+    public bool DiscardMinMaxValues { get; set; }
+    
+    [Orleans.Id(6)]
+    public bool AllowMeasurementsThroughGateway { get; set; }
+    
+    [Orleans.Id(7)]
+    public bool IncludeInDashboard { get; set; }
+    
+    [Orleans.Id(8)]
+    public IDictionary<string, AlertThresholds> AlertRules { get; set; }
 }
