@@ -3,29 +3,28 @@ using bluez.DBus;
 using Microsoft.Extensions.Logging;
 using Orleans;
 
-namespace net.jommy.RuuviCore.GrainServices
+namespace net.jommy.RuuviCore.GrainServices;
+
+public class DeviceListenerFactory
 {
-    public class DeviceListenerFactory
+    private readonly IGrainFactory _grainFactory;
+    private readonly ILoggerFactory _loggerFactory;
+    private const ushort RuuviManufacturerKey = 1177;
+
+    public DeviceListenerFactory(IGrainFactory grainFactory, ILoggerFactory loggerFactory)
     {
-        private readonly IGrainFactory _grainFactory;
-        private readonly ILoggerFactory _loggerFactory;
-        private const ushort RuuviManufacturerKey = 1177;
+        _grainFactory = grainFactory;
+        _loggerFactory = loggerFactory;
+    }
 
-        public DeviceListenerFactory(IGrainFactory grainFactory, ILoggerFactory loggerFactory)
+    public bool TryConstructDeviceListener(IDevice1 device, string deviceAddress, IDictionary<ushort, object> manufacturerData, out IDeviceListener deviceListener)
+    {
+        deviceListener = null;
+        if (manufacturerData.ContainsKey(RuuviManufacturerKey))
         {
-            _grainFactory = grainFactory;
-            _loggerFactory = loggerFactory;
+            deviceListener = new RuuviTagListener(device, deviceAddress, _grainFactory, _loggerFactory.CreateLogger<RuuviTagListener>());
         }
 
-        public bool TryConstructDeviceListener(IDevice1 device, string deviceAddress, IDictionary<ushort, object> manufacturerData, out IDeviceListener deviceListener)
-        {
-            deviceListener = null;
-            if (manufacturerData.ContainsKey(RuuviManufacturerKey))
-            {
-                deviceListener = new RuuviTagListener(device, deviceAddress, _grainFactory, _loggerFactory.CreateLogger<RuuviTagListener>());
-            }
-
-            return deviceListener != null;
-        }
+        return deviceListener != null;
     }
 }
