@@ -2,14 +2,13 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace net.jommy.RuuviCore.GrainServices
+namespace net.jommy.RuuviCore.GrainServices;
+
+public interface IDeviceListener : IDisposable
 {
-    public interface IDeviceListener : IDisposable
-    {
-        Task StartListeningAsync();
+    Task StartListeningAsync();
 
-        bool IsAlive();
+    bool IsAlive();
 
-        Task HandleDataAsync(IDictionary<ushort, object> manufacturerData);
-    }
+    Task HandleDataAsync(IDictionary<ushort, object> manufacturerData);
 }
