@@ -1,10 +1,9 @@
 using System;
 using net.jommy.RuuviCore.Interfaces;
 
-namespace net.jommy.RuuviCore.Grains.DataParsers
+namespace net.jommy.RuuviCore.Grains.DataParsers;
+
+public interface IRuuviDataParser
 {
-    public interface IRuuviDataParser
-    {
-        bool TryParseMeasurements(ReadOnlySpan<byte> data, bool validateValues, out Measurements measurements);
-    }
+    bool TryParseMeasurements(ReadOnlySpan<byte> data, bool validateValues, out MeasurementDTO measurements);
 }
