@@ -3,78 +3,76 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using net.jommy.RuuviCore.Common;
 using net.jommy.RuuviCore.Gateway.Models;
 using net.jommy.RuuviCore.Interfaces;
 using Orleans;
 
-namespace net.jommy.RuuviCore.Gateway.Controllers
+namespace net.jommy.RuuviCore.Gateway.Controllers;
+
+[ApiController]
+[Route("api/v1/ruuvigateway")]
+public class RuuviGatewayController : ControllerBase
 {
-    [ApiController]
-    [Route("api/v1/ruuvigateway")]
-    public class RuuviGatewayController : ControllerBase
+    private readonly ILogger<RuuviGatewayController> _logger;
+    private readonly IClusterClient _clusterClient;
+    private readonly RestApiOptions _options;
+
+    public RuuviGatewayController(ILogger<RuuviGatewayController> logger, IOptions<RestApiOptions> options, IClusterClient clusterClient)
     {
-        private readonly ILogger<RuuviGatewayController> _logger;
-        private readonly IClusterClient _clusterClient;
-        private readonly RestApiOptions _options;
+        _logger = logger;
+        _clusterClient = clusterClient;
+        _options = options.Value;
+    }
 
-        public RuuviGatewayController(ILogger<RuuviGatewayController> logger, IOptions<RestApiOptions> options, IClusterClient clusterClient)
+    [HttpGet("")]
+    public IActionResult TestAPI()
+    {
+        return new JsonResult(new
         {
-            _logger = logger;
-            _clusterClient = clusterClient;
-            _options = options.Value;
+            Message = "If you see this message, it means your Ruuvi HTTP gateway is working!",
+            Instruction = "Now just setup the gateway address in your RuuviApp to point to the same url but with /addmeasurements postfix."
+        });
+    }
+
+    [HttpPost("addmeasurements")]
+    public async Task<IActionResult> AddMeasurements([FromBody] RootObject measurementData)
+    {
+        if (!_options.AllowedDeviceIds.Contains(measurementData.deviceId))
+        {
+            _logger.LogWarning("Received measurement data from a device {deviceId} which is not whitelisted.", measurementData.deviceId);
+            return Unauthorized("Device not whitelisted to submit measurement data.");
         }
 
-        [HttpGet("")]
-        public IActionResult TestAPI()
+        var storeTasks = measurementData.tags.Select(StoreMeasurements);
+
+        await Task.WhenAll(storeTasks);
+
+        return new OkResult();
+    }
+
+    private async Task StoreMeasurements(Tag tag)
+    {
+        var ruuviTag = _clusterClient.GetGrain<IRuuviTag>(tag.id);
+        if (await ruuviTag.MeasurementsAllowedThroughGateway())
         {
-            return new JsonResult(new
+            await ruuviTag.StoreMeasurementData(new MeasurementDTO
             {
-                Message = "If you see this message, it means your Ruuvi HTTP gateway is working!",
-                Instruction = "Now just setup the gateway address in your RuuviApp to point to the same url but with /addmeasurements postfix."
-            });
-        }
-
-        [HttpPost("addmeasurements")]
-        public async Task<IActionResult> AddMeasurements([FromBody] RootObject measurementData)
-        {
-            if (!_options.AllowedDeviceIds.Contains(measurementData.deviceId))
-            {
-                _logger.LogWarning("Received measurement data from a device {deviceId} which is not whitelisted.", measurementData.deviceId);
-                return Unauthorized("Device not whitelisted to submit measurement data.");
-            }
-
-            var storeTasks = measurementData.tags.Select(StoreMeasurements);
-
-            await Task.WhenAll(storeTasks);
-
-            return new OkResult();
-        }
-
-        private async Task StoreMeasurements(Tag tag)
-        {
-            var ruuviTag = _clusterClient.GetGrain<IRuuviTag>(tag.id.ToActorGuid());
-            if (await ruuviTag.MeasurementsAllowedThroughGateway())
-            {
-                await ruuviTag.StoreMeasurementData(new Measurements
+                Acceleration = new Acceleration
                 {
-                    Acceleration = new Acceleration
-                    {
-                        XAxis = (decimal) tag.accelX,
-                        YAxis = (decimal) tag.accelY,
-                        ZAxis = (decimal) tag.accelZ
-                    },
-                    Humidity = (decimal) tag.humidity,
-                    Pressure = (decimal) tag.pressure,
-                    Temperature = (decimal) tag.temperature,
-                    Timestamp = tag.updateAt.ToUniversalTime(),
-                    BatteryVoltage = (int) (tag.voltage * 1000), // convert to mV
-                    RSSI = (short) tag.rssi,
-                    MovementCounter = tag.movementCounter,
-                    TransmissionPower = (int) tag.txPower,
-                    SequenceNumber = tag.measurementSequenceNumber.GetValueOrDefault()
-                });
-            }
+                    XAxis = (decimal) tag.accelX,
+                    YAxis = (decimal) tag.accelY,
+                    ZAxis = (decimal) tag.accelZ
+                },
+                Humidity = (decimal) tag.humidity,
+                Pressure = (decimal) tag.pressure,
+                Temperature = (decimal) tag.temperature,
+                Timestamp = tag.updateAt.ToUniversalTime(),
+                BatteryVoltage = (int) (tag.voltage * 1000), // convert to mV
+                RSSI = (short) tag.rssi,
+                MovementCounter = tag.movementCounter,
+                TransmissionPower = (int) tag.txPower,
+                SequenceNumber = tag.measurementSequenceNumber.GetValueOrDefault()
+            });
         }
     }
 }
