@@ -1,14 +1,13 @@
 using System.Collections.Generic;
 
-namespace net.jommy.RuuviCore
-{
-    public class RestApiOptions
-    {
-        public RestApiOptions()
-        {
-            AllowedDeviceIds = new List<string>();
-        }
+namespace net.jommy.RuuviCore;
 
-        public List<string> AllowedDeviceIds { get; set; }
+public class RestApiOptions
+{
+    public RestApiOptions()
+    {
+        AllowedDeviceIds = new List<string>();
     }
+
+    public List<string> AllowedDeviceIds { get; set; }
 }
