@@ -2,10 +2,9 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using net.jommy.RuuviCore.Interfaces;
 
-namespace net.jommy.RuuviCore.Services
+namespace net.jommy.RuuviCore.Services;
+
+public interface IRuuviTagRepository
 {
-    public interface IRuuviTagRepository
-    {
-        Task<List<RuuviTagInfo>> GetRuuviTagsAsync();
-    }
+    Task<List<RuuviTagInfo>> GetRuuviTagsAsync();
 }
